Reorganise delay parameter handling.

+++ arangod/Agency/AgencyFeature.h
@@ -64,8 +64,8 @@ class AgencyFeature : public ArangodFeature {
   uint64_t _maxAppendSize;
   double _supervisionGracePeriod;
   double _supervisionOkThreshold;
-  double _supervisionDelayAddFollower;
-  double _supervisionDelayFailedFollower;
+  uint64_t _supervisionDelayAddFollower;
+  uint64_t _supervisionDelayFailedFollower;
   std::string _agencyMyAddress;
   std::vector<std::string> _agencyEndpoints;
   std::string _recoveryId;
