+++ arangod/Agency/AgentConfiguration.h
@@ -53,8 +53,8 @@ struct config_t {
   uint64_t _compactionKeepSize;
   double _supervisionGracePeriod;
   double _supervisionOkThreshold;
-  double _supervisionDelayAddFollower;
-  double _supervisionDelayFailedFollower;
+  uint64_t _supervisionDelayAddFollower;
+  uint64_t _supervisionDelayFailedFollower;
   size_t _version;
   std::string _startup;
   size_t _maxAppendSize;
@@ -95,7 +95,7 @@ struct config_t {
   config_t(std::string const& rid, size_t as, double minp, double maxp,
            std::string const& e, std::vector<std::string> const& g, bool s,
            bool st, bool w, double f, uint64_t c, uint64_t k, double p,
-           double o, double q, double r, size_t a);
+           double o, uint64_t q, uint64_t r, size_t a);
 
   /// @brief copy constructor
   config_t(config_t const&);
@@ -231,10 +231,10 @@ struct config_t {
   double supervisionOkThreshold() const;
 
   /// @brief Supervision delay add follower
-  double supervisionDelayAddFollower() const;
+  uint64_t supervisionDelayAddFollower() const;
 
   /// @brief Supervision delay failed follower
-  double supervisionDelayFailedFollower() const;
+  uint64_t supervisionDelayFailedFollower() const;
 
   /// @brief set Supervision grace period
   void setSupervisionGracePeriod(double d) { _supervisionGracePeriod = d; }
@@ -243,12 +243,12 @@ struct config_t {
   void setSupervisionOkThreshold(double d) { _supervisionOkThreshold = d; }
 
   /// @brief set Supervision delay add follower
-  void setSupervisionDelayAddFollower(double d) {
+  void setSupervisionDelayAddFollower(uint64_t d) {
     _supervisionDelayAddFollower = d;
   }
 
   /// @brief set Supervision delay failed follower
-  void setSupervisionDelayFailedFollower(double d) {
+  void setSupervisionDelayFailedFollower(uint64_t d) {
     _supervisionDelayFailedFollower = d;
   }
